Extracted get_request.go
